# docs/phase2/openapi-v2.yaml
openapi: 3.0.3
info:
  title: TeamCollab MVP API
  version: 2.0.0
  description: >
    团队协作工具 MVP REST API（看板/任务/成员/评论/通知）。Phase 2 契约，前后端联调唯一依据。

    统一响应格式：{ "code": 0, "data": {}, "message": "" }，code=0 成功。
    错误码：40001 参数校验失败 / 40101 未认证或 token 过期 / 40301 无权限(RBAC) /
    40401 资源不存在(含跨租户访问，不泄露存在性) / 40901 资源冲突 / 42901 频率限制 / 50000 服务器内部错误。

    认证：Authorization: Bearer <access_token>。SSE 端点因 EventSource 无法自定义 Header，走 query token。

    版本锚定：next@15.5.9+ / fastapi@0.140.x / lucide-react@1.24.0（全项目唯一图标库，禁止 emoji 图标）。
servers:
  - url: http://localhost:8000
    description: FastAPI 本地开发
...
paths:
...
  /api/v1/teams:
    post:
      tags: [teams]
      summary: 创建团队（创建者自动成为 owner）
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          description: 参数校验失败（code=40001）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/teams/{teamId}/members:
    get:
      tags: [teams]
      summary: 团队成员列表（member+）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: teamId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: 成员列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberListResponse'
        '403':
          description: 无权限（code=40301）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 团队不存在或跨租户（code=40401）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [teams]
      summary: 邀请成员（admin+，按邮箱匹配已注册用户）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: teamId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email: { type: string, format: email }
                role: { $ref: '#/components/schemas/Role' }
      responses:
        '201':
          description: 已邀请，返回新成员
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
        '400':
          description: 邮箱不存在于平台 / role 非法 / 已是成员（code=40001 或 40901）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限（code=40301）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/teams/{teamId}/members/{userId}/role:
    patch:
      tags: [teams]
      summary: 修改成员角色（admin+；owner 不可被降级、不可被修改）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: teamId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { $ref: '#/components/schemas/Role' }
      responses:
        '200':
          description: 已更新，返回更新后成员
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResponse'
        '400':
          description: role 非法 / 目标是 owner（code=40001）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限（code=40301）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 成员不存在或跨租户（code=40401）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/teams/{teamId}/members/{userId}:
    delete:
      tags: [teams]
      summary: 移除成员（admin+；owner 不可被移除）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: teamId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: 已移除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '403':
          description: 无权限或目标是 owner（code=40301）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 成员不存在或跨租户（code=40401）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------- BOARDS ----------------
  /api/v1/teams/{teamId}/boards:
    get:
      tags: [boards]
      summary: 看板列表（member+）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: teamId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: 看板列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardListResponse'
        '403':
          description: 无权限（code=40301）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [boards]
      summary: 创建看板（member+）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: teamId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardResponse'
        '400':
          description: 参数校验失败（code=40001）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/boards/{boardId}:
    get:
      tags: [boards]
      summary: 看板详情（含列 + 各列任务，按 position 排序）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: boardId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: 看板详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardDetailResponse'
        '404':
          description: 看板不存在或跨租户（code=40401）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [boards]
      summary: 更新看板（member+，仅 name）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: boardId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardResponse'
        '403':
          description: 无权限（code=40301）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 看板不存在或跨租户（code=40401）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [boards]
      summary: 删除看板（admin+，级联删除列与任务）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: boardId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: 已删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '403':
          description: 无权限（code=40301）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 看板不存在或跨租户（code=40401）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/boards/{boardId}/columns:
    post:
      tags: [boards]
      summary: 创建列（member+）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: boardId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
                position: { type: integer, default: 0, description: 缺省追加到末尾（当前最大 position + 1） }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColumnResponse'
        '400':
          description: 参数校验失败（code=40001）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 看板不存在或跨租户（code=40401）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/columns/{columnId}:
    patch:
      tags: [boards]
      summary: 更新列（member+，name/position）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: columnId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
                position: { type: integer }
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColumnResponse'
        '403':
          description: 无权限（code=40301）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 列不存在或跨租户（code=40401）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------- TASKS ----------------
...
  /api/v1/tasks/{taskId}/comments:
    get:
      tags: [comments]
      summary: 评论列表（member+，按 created_at 升序分页）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: page, in: query, required: false, schema: { type: integer, default: 1, minimum: 1 } }
        - { name: limit, in: query, required: false, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200':
          description: 分页评论列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedComments'
        '404':
          description: 任务不存在或跨租户（code=40401）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [comments]
      summary: 发评论（member+，检测 @username 提及并通知被提及成员）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, minLength: 1, maxLength: 2000, description: 支持 @username 提及，格式 @后跟用户 display_name 或邮箱前缀 }
      responses:
        '201':
          description: 创建成功（若命中 @提及，已向被提及者生成 comment_added 通知）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '400':
          description: 参数校验失败（code=40001）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 任务不存在或跨租户（code=40401）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------- NOTIFICATIONS ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, data, message]
      properties:
        code:
          type: integer
          enum: [40001, 40101, 40301, 40401, 40901, 42901, 50000]
          description: 非 0 即错误
        data:
          type: object
          nullable: true
          description: 错误时恒为 null
        message:
          type: string
          description: 人类可读错误描述（不泄露内部细节）
    OkResponse:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer, example: 0 }
        data:
          type: object
          properties:
            ok: { type: boolean, example: true }
        message: { type: string, example: '' }
...
    Role:
      type: string
      enum: [owner, admin, member, viewer]
      description: owner=团队创建者(不可降级) / admin=管理员 / member=成员 / viewer=只读
    Priority:
      type: string
      enum: [low, medium, high, urgent]
    TaskStatus:
      type: string
      enum: [todo, in_progress, done]
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        created_at: { type: string, format: date-time }
    Team:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        created_at: { type: string, format: date-time }
    TeamResponse:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer, example: 0 }
        data: { $ref: '#/components/schemas/Team' }
        message: { type: string, example: '' }
    Member:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
        role: { $ref: '#/components/schemas/Role' }
        created_at: { type: string, format: date-time }
    MemberResponse:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer, example: 0 }
        data: { $ref: '#/components/schemas/Member' }
        message: { type: string, example: '' }
    MemberListResponse:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer, example: 0 }
        data:
          type: array
          items: { $ref: '#/components/schemas/Member' }
        message: { type: string, example: '' }

    # ---------------- 看板/列 ----------------
    Board:
      type: object
      properties:
        id: { type: string, format: uuid }
        team_id: { type: string, format: uuid }
        name: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    BoardResponse:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer, example: 0 }
        data: { $ref: '#/components/schemas/Board' }
        message: { type: string, example: '' }
    BoardListResponse:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer, example: 0 }
        data:
          type: array
          items: { $ref: '#/components/schemas/Board' }
        message: { type: string, example: '' }
    BoardColumn:
      type: object
      properties:
        id: { type: string, format: uuid }
        board_id: { type: string, format: uuid }
        name: { type: string }
        position: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ColumnResponse:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer, example: 0 }
        data: { $ref: '#/components/schemas/BoardColumn' }
        message: { type: string, example: '' }
    BoardColumnWithTasks:
      type: object
      properties:
        id: { type: string, format: uuid }
        board_id: { type: string, format: uuid }
        name: { type: string }
        position: { type: integer }
        tasks:
          type: array
          items: { $ref: '#/components/schemas/Task' }
    BoardDetailResponse:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer, example: 0 }
        data:
          type: object
          properties:
            id: { type: string, format: uuid }
            team_id: { type: string, format: uuid }
            name: { type: string }
            columns:
              type: array
              items: { $ref: '#/components/schemas/BoardColumnWithTasks' }
        message: { type: string, example: '' }

    # ---------------- 任务 ----------------
...
    Task:
      type: object
      properties:
        id: { type: string, format: uuid }
        board_id: { type: string, format: uuid }
        board_name: { type: string, description: 我的任务聚合时附带，便于跳转 }
        column_id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string, nullable: true }
        assignee_id: { type: string, format: uuid, nullable: true }
        assignee:
          type: object
          nullable: true
          properties:
            id: { type: string, format: uuid }
            display_name: { type: string }
        priority: { $ref: '#/components/schemas/Priority' }
        status: { $ref: '#/components/schemas/TaskStatus' }
        position: { type: integer }
        due_date: { type: string, format: date, nullable: true }
        start_date: { type: string, format: date, nullable: true }
        created_by: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Comment:
      type: object
      properties:
        id: { type: string, format: uuid }
        task_id: { type: string, format: uuid }
        author:
          type: object
          properties:
            id: { type: string, format: uuid }
            display_name: { type: string }
        content: { type: string, maxLength: 2000 }
        mentions:
          type: array
          items: { type: string, format: uuid }
          description: 被 @ 提及的用户 ID 列表（检测结果，服务端返回）
        created_at: { type: string, format: date-time }
    CommentResponse:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer, example: 0 }
        data: { $ref: '#/components/schemas/Comment' }
        message: { type: string, example: '' }
    PaginatedComments:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer, example: 0 }
        data:
          type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Comment' }
            total: { type: integer }
            page: { type: integer }
            limit: { type: integer }
            hasMore: { type: boolean }
        message: { type: string, example: '' }

    # ---------------- 通知 ----------------
...